Initialize axios for GitHub GraphQL
We passed GitHub's GraphQL endpoint (URL) to axios and configured it to
use an HTTP header named "Authorization" with the value:

    bearer YOUR_GITHUB_ACCESS_TOKEN

Since the access token is secret information, we wouldn't want it to be
leaked by hard coding it into source code which is checked into a
repository. Hence, we configured it to be read from the environment
variables.

To make defining the access token in the environment variables of the
development machine simple, we created a file named ".env" in the root
of the source code. However, we DID NOT check this into the version
control, because checking this in would have been the same thing as
simply hard coding the access token. That is, it would have been the
same thing as exposing a secret, where the secret is the access token.