First attempt to use configurable arguments
Previously, the GitHub organization and repository were hard coded to
GraphQL query. This is the first attempt to make them configurable by
converting the query from a JavaScript template string to a JavaScript
function which accepts two arguments (for the organization and
repository) and returns a template string representing the GraphQL query
for the given GitHub organization and repository. Then, use this
function (that constructs the GraphQL query) in "onFetchFromGitHub"
function in order to get the list of issues for a given repository (of a
given organization).