Implement removing star from a repository
Such that the same button is used for both starring and unstarring a
repository. If the repo was starred, the button says "Unstar" and
clicking on it unstars the repo. Whereas if the repo was not starred,
the button says "Star" and clicking on it stars the repo. The button
changes to the other state (from "Star" to "Unstar" or from "Unstar" to
"Star") each time the operation is successfully completed (that is, a
valid response from GitHub's GraphQL server is received).